+++ src/Main.java
@@ -1,0 +1,10 @@
+public class Main {
+
+    public static void main(String[] args) {
+        UI conga = new UI();
+        RandomAgent randomAgent = new RandomAgent(conga);
+        while(true) {
+            randomAgent.move();
+        }
+    }
+}
